+++ lerna.json
@@ -1,0 +1,6 @@
+{
+  "npmClient": "yarn",
+  "packages": ["packages/*"],
+  "version": "5.0.0",
+  "useWorkspaces": true
+}
