+++ packages/core/package.json
@@ -1,0 +1,52 @@
+{
+  "name": "@squiss/core",
+  "version": "5.0.0",
+  "description": "High-volume SQS poller - core",
+  "main": "dist/index.js",
+  "types": "dist/index.d.ts",
+  "engines": {
+    "node": "^6 || ^8 || ^10 || ^12 || ^14 || ^15"
+  },
+  "scripts": {
+    "postinstall": "node -e 'const semver = require(\"semver\");process.exit(semver.lt(process.version, \"10.16.0\"));' && (npm install --no-save @aws-sdk/client-sqs@^3.8.1 @aws-sdk/client-s3@^3.8.1) || (npm install --no-save iltorb@2.4.3 aws-sdk@^2.864.0)",
+    "compile": "tsc",
+    "clean": "rm -rf dist"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/PruvoNet/squiss-ts.git"
+  },
+  "files": [
+    "/dist/*.d.ts",
+    "/dist/*.js"
+  ],
+  "keywords": [
+    "aws",
+    "amazon",
+    "sqs",
+    "polling",
+    "poll",
+    "poller",
+    "queue",
+    "typescript",
+    "sns",
+    "s3",
+    "extended",
+    "client",
+    "sdk"
+  ],
+  "author": "Regev Brody <regevbr@gmail.com>",
+  "license": "Apache-2.0",
+  "bugs": {
+    "url": "https://github.com/PruvoNet/squiss-ts/issues"
+  },
+  "homepage": "https://github.com/PruvoNet/squiss-ts#readme",
+  "dependencies": {
+    "linked-list": "^2.1.0",
+    "semver": "^7.3.4",
+    "ts-type-guards": "^0.7.0",
+    "uuid": "^8.3.2"
+  },
+  "devDependencies": {
+  }
+}
