+++ packages/core/src/TimeoutExtender.ts
@@ -1,4 +1,3 @@
-'use strict';
 
 import {Message} from './Message';
 import * as LinkedList from 'linked-list';
