+++ packages/core/src/Utils.ts
@@ -1,4 +1,3 @@
-'use strict';
 
 export const removeEmptyKeys = <T extends object>(obj: T): T => {
     (Object.keys(obj) as (keyof typeof obj)[]).forEach((key) => {
