+++ packages/core/src/s3Utils.ts
@@ -1,4 +1,3 @@
-'use strict';
 
 import {v4 as uuidv4} from 'uuid';
 import {getSizeInBytes} from './messageSizeUtils';
