+++ packages/core/src/test/stubs/S3Stub.ts
@@ -1,4 +1,3 @@
-'use strict';
 
 import {EventEmitter} from 'events';
 import {
