+++ packages/core/src/test/stubs/SQSStub.ts
@@ -1,4 +1,3 @@
-'use strict';
 
 import {EventEmitter} from 'events';
 import {
